chore: E2E 테스트 컨트롤러 이동
