+++ src/test/java/kr/hhplus/be/server/interfaces/ConcertControllerIntegrationTest.java
@@ -1,11 +1,9 @@
-package kr.hhplus.be.server.interfaces.interceptor.integration;
+package kr.hhplus.be.server.interfaces;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.hhplus.be.server.domain.entity.*;
 import kr.hhplus.be.server.domain.service.QueueService;
 import kr.hhplus.be.server.infra.repository.impl.ReservationRepositoryImpl;
 import kr.hhplus.be.server.infra.repository.jpa.*;
-import kr.hhplus.be.server.interfaces.dto.reservation.ReservationHttpDto;
 import kr.hhplus.be.server.support.type.ConcertStatus;
 import kr.hhplus.be.server.support.type.QueueStatus;
 import kr.hhplus.be.server.support.type.ReservationStatus;
@@ -16,21 +14,22 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.LocalDateTime;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class ReservationControllerIntegrationTest {
+public class ConcertControllerIntegrationTest {
+
     private Queue queue;
     private Queue expiredQueue;
-    private ReservationHttpDto.ReservationRequest request;
+    private Concert concert;
+    private ConcertSchedule concertSchedule;
 
 
     @Autowired
@@ -48,9 +47,11 @@ public class ReservationControllerIntegrationTest {
     @Autowired
     private SeatJpaRepository seatJpaRepository;
     @Autowired
+    private ReservationRepositoryImpl reservationRepository;
+    @Autowired
     private QueueJpaRepository queueJpaRepository;
     @Autowired
-    private ObjectMapper objectMapper; // SpringBootTest에서 자동 주입
+    private QueueService queueService;
 
     @BeforeEach
     void setUp() {
@@ -90,7 +91,7 @@ public class ReservationControllerIntegrationTest {
 
         pointJpaRepository.save(point);
 
-        Concert concert = Concert.builder()
+        concert = Concert.builder()
                 .title("콘서트")
                 .description("콘서트내용")
                 .status(ConcertStatus.OPEN)
@@ -98,7 +99,7 @@ public class ReservationControllerIntegrationTest {
 
         concertJpaRepository.save(concert);
 
-        ConcertSchedule concertSchedule = ConcertSchedule.builder()
+        concertSchedule = ConcertSchedule.builder()
                 .concertId(concert.getId())
                 .availableReservationTime(LocalDateTime.now().minusDays(5))
                 .concertTime(LocalDateTime.now().plusDays(30))
@@ -116,51 +117,111 @@ public class ReservationControllerIntegrationTest {
 
         seatJpaRepository.save(seat);
 
-        request = ReservationHttpDto.ReservationRequest.builder()
+        Reservation reservation = Reservation.builder()
+                .concertScheduleId(concertSchedule.getId())
                 .userId(user.getId())
                 .concertId(concert.getId())
-                .concertScheduleId(concertSchedule.getId())
                 .seatId(seat.getId())
+                .status(ReservationStatus.PAYMENT_WAITING)
+                .reservedAt(LocalDateTime.now())
                 .build();
 
+        reservationRepository.save(reservation);
     }
 
     @Test
-    void 정상_토큰인_경우_예약_성공() throws Exception {
+    void 정상_토큰인_경우_콘서트_조회_성공() throws Exception {
         //given
 
         //when//then
-        mvc.perform(post("/api/reservations")
-                        .header("Token", queue.getToken())
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
-                .andExpect(status().isCreated());
+        mvc.perform(get("/api/concerts")
+                        .header("Token", queue.getToken()))
+                .andExpect(status().isOk());
     }
 
     @Test
-    void 헤더에_토큰이_없는_경우_예약_실패() throws Exception {
+    void 헤더에_토큰이_없는_경우_콘서트_조회_실패() throws Exception {
         //given
 
         //when //then
-        mvc.perform(post("/api/reservations")
-                        .content(objectMapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+        mvc.perform(get("/api/concerts"))
                 .andExpect(status().isUnauthorized())
                 .andExpect(jsonPath("$.message").value("없는 토큰 입니다."));
     }
 
     @Test
-    void 헤더에_토큰이_만료된_경우_예약_실패() throws Exception {
+    void 헤더에_토큰이_만료된_경우_콘서트_조회_실패() throws Exception {
         //given
 
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime expiredTime = expiredQueue.getExpiredAt();
+
         //when //then
-        mvc.perform(post("/api/reservations")
-                        .header("Token", expiredQueue.getToken())
-                        .content(objectMapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+        mvc.perform(get("/api/concerts")
+                        .header("Token", expiredQueue.getToken()))
                 .andExpect(status().isUnauthorized())
                 .andExpect(jsonPath("$.message").value("유효하지 않는 토큰 입니다."));
     }
 
-}
+    @Test
+    void 정상_토큰인_경우_콘서트스케쥴_조회_성공() throws Exception {
+        //given
 
+        //when//then
+        mvc.perform(get("/api/concerts/{concertId}/schedules", concert.getId())
+                        .header("Token", queue.getToken()))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    void 헤더에_토큰이_없는_경우_콘서트스케쥴_조회_실패() throws Exception {
+        //given
+
+        //when //then
+        mvc.perform(get("/api/concerts/{concertId}/schedules", concert.getId()))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.message").value("없는 토큰 입니다."));
+    }
+
+    @Test
+    void 헤더에_토큰이_만료된_경우_콘서트스케쥴_조회_실패() throws Exception {
+        //given
+
+        //when //then
+        mvc.perform(get("/api/concerts/{concertId}/schedules", concert.getId())
+                        .header("Token", expiredQueue.getToken()))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.message").value("유효하지 않는 토큰 입니다."));
+    }
+
+    @Test
+    void 정상_토큰인_경우_좌석들_조회_성공() throws Exception {
+        //given
+
+        //when//then
+        mvc.perform(get("/api/concerts/{concertId}/schedules/{concertScheduleId}/seats", concert.getId(), concertSchedule.getId())
+                        .header("Token", queue.getToken()))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    void 헤더에_토큰이_없는_경우_좌석들_조회_실패() throws Exception {
+        //given
+
+        //when //then
+        mvc.perform(get("/api/concerts/{concertId}/schedules/{concertScheduleId}/seats", concert.getId(), concertSchedule.getId()))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.message").value("없는 토큰 입니다."));
+    }
+
+    @Test
+    void 헤더에_토큰이_만료된_경우_좌석들_조회_실패() throws Exception {
+        //given
+
+        //when //then
+        mvc.perform(get("/api/concerts/{concertId}/schedules/{concertScheduleId}/seats", concert.getId(), concertSchedule.getId())
+                        .header("Token", expiredQueue.getToken()))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.message").value("유효하지 않는 토큰 입니다."));
+    }
+}
